TextClassification/flask/requirements.txt
absl-py==0.11.0
astunparse==1.6.3
async-generator==1.10
backports.functools-lru-cache==1.6.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
Flask==1.1.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
ipython==5.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandocfilters==1.4.2
prompt-toolkit==1.0.15
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
Send2Trash==1.5.0
simplegeneric==0.8.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
testpath==0.4.4
urllib3==1.25.11
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
